chore: add added block state to drawer context
