+++ apps/studio/src/features/editing-experience/components/ComplexEditorStateDrawer.tsx
@@ -3,8 +3,11 @@ import { IconButton } from "@opengovsg/design-system-react"
 import { BiDollar, BiX } from "react-icons/bi"
 
 import FormBuilder from "./form-builder/FormBuilder"
+import { useEditorDrawerContext } from "~/contexts/EditorDrawerContext"
 
 export default function ComplexEditorStateDrawer(): JSX.Element {
+  const { addedBlock: component } = useEditorDrawerContext()
+
   return (
     <Box position="relative" h="100%" w="100%">
       <Box
