+++ apps/studio/src/features/editing-experience/components/EditPageDrawer.tsx
@@ -3,9 +3,7 @@ import ComponentSelector from '~/components/PageEditor/ComponentSelector'
 import RootStateDrawer from './RootStateDrawer'
 import TipTapComponent from './TipTapComponent'
 import ComplexEditorStateDrawer from './ComplexEditorStateDrawer'
-import { IsomerNativeComponentsMap } from '@opengovsg/isomer-components'
-import { TypeCompiler } from '@sinclair/typebox/compiler'
-import type { ProseProps } from '@opengovsg/isomer-components/dist/cjs/interfaces'
+import { inferAsProse } from '../utils'
 
 
 export function EditPageDrawer() {
@@ -14,21 +12,6 @@ export function EditPageDrawer() {
     drawerState: currState,
     currActiveIdx,
   } = useEditorDrawerContext()
-
-  const proseSchema = IsomerNativeComponentsMap.prose
-  const compiled = TypeCompiler.Compile(proseSchema)
-
-  const inferAsProse = (component?: typeof pageState[number]): ProseProps => {
-    if (!component) {
-      throw new Error(`Expected component of type prose but got undefined`)
-    }
-
-    if (compiled.Check(component)) {
-      return component
-    }
-
-    throw new Error(`Expected component of type prose but got type ${component.type}`)
-  }
 
   switch (currState.state) {
     case "root":
