+++ apps/studio/src/features/editing-experience/components/form-builder/FormBuilder.tsx
@@ -68,7 +68,7 @@ export default function FormBuilder({
 
   return (
     <JsonForms
-      schema={subSchema || {}}
+      schema={subSchema}
       data={formData}
       renderers={renderers}
       onChange={({ data }) => {
